Add files via upload
Here for the exercise classification problem, prediction model is created with random forest. It is found that identifying the most influential parameters by creating a subset of training data and using only those parameters to create a prediction model using the entire training data gives good time saving with very little loss of accuracy. In this case for a sacrifice of ~2% accuracy the prediction models could be made in ~1/5th of the time. Request reviews.